palette editor: editor switch to edit mode, when clicked on palette tile

+++ hswidget/palettegridwidget/state.go
@@ -31,7 +31,7 @@ func (ws *widgetState) Dispose() {
 
 type editEntryState struct {
 	idx     int
-	r, g, b int32
+	r, g, b uint8
 }
 
 func (ees *editEntryState) Dispose() {
